requirements.txt
RPi.GPIO
smbus-cffi
paho-mqtt
Adafruit_MPR121